+++ ec-project-api/Dtos/response/system/StatusDto.cs
@@ -2,7 +2,7 @@ namespace ec_project_api.Dtos.Statuses
 {
     public class StatusDto
     {
-        public int StatusId { get; set; }
+        public short StatusId { get; set; }
         public string Name { get; set; } = null!;
         public string? DisplayName { get; set; }
         public string EntityType { get; set; } = null!;
